refactor: pull out xsalsa20 cipher tests
