+++ example/src/hooks/useTestsList.ts
@@ -4,6 +4,7 @@ import { TestsContext } from '../tests/util';
 
 import '../tests/cipher/cipher_tests';
 import '../tests/cipher/chacha_tests';
+import '../tests/cipher/xsalsa20_tests';
 import '../tests/ed25519/ed25519_tests';
 import '../tests/hash/hash_tests';
 import '../tests/hmac/hmac_tests';
