+++ example/src/tests/cipher/cipher_tests.ts
@@ -3,7 +3,6 @@ import {
   getCiphers,
   createCipheriv,
   randomFillSync,
-  xsalsa20,
 } from 'react-native-quick-crypto';
 import { expect } from 'chai';
 import { test } from '../util';
@@ -13,10 +12,6 @@ const SUITE = 'cipher';
 
 // --- Constants and Test Data ---
 const key16 = Buffer.from('a8a7d6a5d4a3d2a1a09f9e9d9c8b8a89', 'hex');
-const key32 = Buffer.from(
-  'a8a7d6a5d4a3d2a1a09f9e9d9c8b8a89a8a7d6a5d4a3d2a1a09f9e9d9c8b8a89',
-  'hex',
-);
 const iv16 = randomFillSync(new Uint8Array(16));
 const iv12 = randomFillSync(new Uint8Array(12)); // Common IV size for GCM/CCM/OCB
 const iv = Buffer.from(iv16);
@@ -110,17 +105,4 @@ allCiphers.forEach(cipherName => {
       expect.fail(`Cipher ${cipherName} threw an error: ${message}`);
     }
   });
-});
-
-// libsodium cipher tests
-test(SUITE, 'xsalsa20', () => {
-  const key = new Uint8Array(key32);
-  const nonce = randomFillSync(new Uint8Array(24));
-  const data = new Uint8Array(plaintextBuffer);
-  // encrypt
-  const ciphertext = xsalsa20(key, nonce, data);
-  // decrypt - must use the same nonce as encryption
-  const decrypted = xsalsa20(key, nonce, ciphertext);
-  // test decrypted == data
-  expect(decrypted).eql(data);
 });
